requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
openai==1.55.3
pydantic==2.10.3
python-dotenv==1.0.1
jinja2==3.1.4

python-telegram-bot==22.5
pandas==2.2.3
numpy==2.1.3
